backend/requirements.txt
asgiref==3.4.1
cffi==1.15.0
cryptography==36.0.0
Django==3.2.9
django-cors-headers==3.13.0
django-environ==0.9.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.0
loguru==0.5.3
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
pytz==2021.3
sqlparse==0.4.2
